adding jwt auth

+++ backend/routes/expenses.ts
@@ -1,23 +1,40 @@
 import { Request, Response } from "express";
 import IExpense from "../interfaces/expense";
 const express = require("express");
+const { User, Expense } = require("../schemas");
+const { getUser } = require("../jwt")
 const expensesRouter = express.Router();
 
-const { User, Expense } = require("../schemas");
 
 expensesRouter.get("/:userId", async (req: Request, res: Response) => {
 
   try {
     const userId = req.params.userId;
 
+    const token = req.headers['authorization']?.split(" ")[1]
+    const authUser = getUser(token);
+
+    if (userId != authUser?.id && authUser?.role !== "admin") {
+      return res.status(403).send({
+        error: "You are not authorized",
+        message: "You are not authorized"
+      });
+    }
+
     const user = await User.findById(userId);
     if (!user) {
-      return res.status(404).send("User not found");
+      return res.status(404).send({
+        error: "User not found",
+        message: "User not found"
+      });
     }
 
     const expenses = await Expense.find({ user: userId })
     if (!expenses) {
-      return res.status(404).send("Expenses not found");
+      return res.status(404).send({
+        error: "Expenses not found",
+        message: "Expenses not found"
+      });
     }
 
     return res.json(expenses);
@@ -32,11 +49,23 @@ expensesRouter.post("/", async (req: Request, res: Response) => {
   try {
     const expense: IExpense = req.body;
 
+    const token = req.headers['authorization']?.split(" ")[1]
+    const authUser = getUser(token);
+
+    if (expense.user != authUser?.id && authUser?.role !== "admin") {
+      return res.status(403).send({
+        error: "You are not authorized",
+        message: "You are not authorized"
+      });
+    }
     // validate user
     const user = await User.findById(expense.user);
 
     if (!user) {
-      return res.status(400).send("User not found");
+      return res.status(400).send({
+        error: "User not found",
+        message: "User not found",
+      });
     }
 
     const newExpense = await Expense.create(expense);
@@ -57,11 +86,23 @@ expensesRouter.put("/:expenseId", async (req: Request, res: Response) => {
     const expense: IExpense = req.body;
 
 
+    const token = req.headers['authorization']?.split(" ")[1]
+    const authUser = getUser(token);
+
+    if (expense.user != authUser?.id && authUser?.role !== "admin") {
+      return res.status(403).send({
+        error: "You are not authorized",
+        message: "You are not authorized"
+      });
+    }
     const updatedExpense = await Expense.findByIdAndUpdate(expenseId, expense, {
       new: true,
     });
     if (!updatedExpense) {
-      return res.status(404).send("Expense not found");
+      return res.status(404).send({
+        error: "Expense not found",
+        message: "Expense not found",
+      });
     }
     return res.status(201).json(updatedExpense);
 
@@ -78,9 +119,22 @@ expensesRouter.put("/:expenseId", async (req: Request, res: Response) => {
 expensesRouter.delete("/:expenseId", async (req: Request, res: Response) => {
   try {
     const expenseId = req.params.expenseId;
+    const expense = await Expense.findById(expenseId);
+    const token = req.headers['authorization']?.split(" ")[1]
+    const authUser = getUser(token);
+
+    if (expense.user != authUser?.id && authUser?.role !== "admin") {
+      return res.status(403).send({
+        error: "You are not authorized",
+        message: "You are not authorized"
+      });
+    }
     const deletedExpense = await Expense.findByIdAndDelete(expenseId);
     if (!deletedExpense) {
-      return res.status(404).send("Expense not found");
+      return res.status(404).send({
+        error: "Expense not found",
+        message: "Expense not found",
+      });
     }
     return res.status(201).json(deletedExpense);
   } catch (err) {
